LPS-94365 Move test classes tracking to ClientState
